+++ launchelf.h
@@ -342,7 +342,7 @@ int keyboard(char *out, int max);
 void genLimObjName(char *uLE_path, int reserve);
 int genFixPath(const char *inp_path, char *gen_path);
 int genOpen(char *path, int mode);
-int genLseek(int fd, int where, int how);
+s64 genLseek(int fd, s64 where, int how);
 int genRead(int fd, void *buf, int size);
 int genWrite(int fd, void *buf, int size);
 int genClose(int fd);
